requirements.txt
Django==3.2.4
celery==5.1.2
redis==3.5.3
psycopg2-binary==2.9.1
feedparser==6.0.8
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
pyyaml==5.4.1
uritemplate==3.0.1